requirements.txt
# Production dependencies (Railway deploy).
# Heavy research-only deps (transformers, torch, pytrends) moved to
# requirements-research.txt to keep the API container lean.

# Core data + engine
yfinance>=0.2.31
pandas>=2.0
numpy>=1.24
requests>=2.28
python-dotenv>=1.0
scipy>=1.11

# Nightly self-learning (NN-backed asymmetric scorer + confidence model)
scikit-learn>=1.3

# API server
fastapi>=0.100
uvicorn[standard]>=0.23

# Auth + billing
stripe>=10.0
PyJWT[crypto]>=2.8
httpx>=0.27
slowapi>=0.1.9
